+++ CHANGELOG.md
@@ -9,19 +9,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- SDK (**breaking change**): Added `TimelineViewModel.onProcessedScrollTo`, which should be called, when the UI fully finished processing `TimelineViewModel.scrollTo`
+- UI,SDK,WEB: Play audio on web (enabled by default, can be disabled with existing feature toggle `media player`)
 - UI: Add proper rendering for "All accounts" avatar
 - SDK: Ability for rovingFocusContainer to scroll to an item before focusing it
 - SDK: Ability for rovingFocusContainer to change focused item for lazy loading containers
 - CI, INTERNAL: Add and enforce ktfmt formatting
 - SDK,UI: Add delete avatar functionality
 - SDK: Sending audio files now sets `info.duration` if possible
-- SDK,UI: Add new search API to include search providers (e.g., an LDAP search) to look for when searching users
+- UI,SDK,ANDROID: Drafts for voice messages
+- SDK,UI: Add new search API to include search providers (e.g., an LDAP search) to look for when searching users (see README.md for more details)
 
 ### Changed
 
 - SDK: (**breaking**) moved `FileKitFileDescriptor` from `de.connect2x.trixnity.messenger.compose.view.files` to
   `de.connect2x.trixnity.messenger.util`
-- CI: Use shell runner for linux
+- DEPENDENCY: (**breaking**) update Lognity to 2.1.0
+- CI: Use shell runner for linux (two different runners depending on workload)
 
 ### Deprecated
 
@@ -29,7 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- UI: Multiple timeline-jump issues
 - UI: Some containers not being focusable, when selected item was out of view
+- UI: Draft not showing in roomlist when room empty
 - SDK: Slow compilation times
 - UI,A11y: Some containers not being focusable, when selected item was out of view
 - SDK: Add error handling for player on iOS and Android
@@ -37,6 +43,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UI,SDK: It is no longer possible to create encrypted public rooms
 - UI,SDK: It is now possible to create unencrypted private rooms
 - SDK: Failure on first login under Windows
+- SDK: Shortly wait for decrypting before show wait message
+- UI,SDK: Wrong text when receiving member state events with no change
+- SDK: Voice messages are saved with correct file extension
 
 ### Security